Controller/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
cryptography==46.0.3
deprecation==2.1.0
fastavro==1.12.1
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.2.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
postgrest==2.25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.25.0
requests==2.32.5
schedule==1.2.2
shellingham==1.5.4
six==1.17.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
websockets==15.0.1
yarl==1.22.0
